+++ packages/harmony/src/components/comments/Timestamp/index.ts
@@ -1,2 +1,3 @@
 export { Timestamp } from './Timestamp'
+export { getLargestTimeUnitText } from './util'
 export * from './types'
